Split out Exporter class
